Add missing json tags for oneof.
According to this:

https://github.com/golang/protobuf/issues/446

This isn't a bug, and yet I agree with the last reply that was ignored
asking why the tags are provided elsewhere. Seems to be that those tags
should be removed if no effort is going to be made to keep them
consistent with each other.

Frustrating. A good argument could be made here that we should have
an automated test that checks JSON output, to guard against a regression.